add verticalDivider
