--- src/DividerVertical/index.ts
@@ -1,1 +1,0 @@
-export { default as VerticalDivider } from "./DividerVertical";
