+++ src/VerticalDivider/index.ts
@@ -1,0 +1,1 @@
+export { default as VerticalDivider } from "./VerticalDivider";
